+++ CHANGELOG.rst
@@ -28,6 +28,11 @@ not released yet
 * NEW khal learned the ``--logfile/-l LOGFILE`` flag which allows logging to a
   file
 * NEW format can now print the duration of an event with `{duration}`
+* NEW configuration option: [view]monthdisplay = firstday|plainweek,
+  if set to 'firstday', khal displays the month name as soon as any day
+  in the week is within the new month. If set to 'plainweek', khal
+  displays the month name only if the first day of the week is winth
+  the new month.
 
 0.9.6
 =====